+++ src/Domain/SelectionEngine/ISelectionEngine.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using Domain;
 
 namespace Domain.Selection;
@@ -11,6 +12,12 @@ public interface ISelectionEngine
     IReadOnlyList<SelectionHistoryEntry> GetHistory();
 
     void ClearHistory();
+
+    GameUserData GetUserData(uint appId);
+
+    IReadOnlyDictionary<uint, GameUserData> GetUserDataSnapshot();
+
+    void UpdateUserData(uint appId, GameUserData userData);
 
     GameEntry PickNext(IEnumerable<GameEntry> games);