+++ src/Domain/SelectionEngine/SelectionSettings.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace Domain.Selection;
 
 internal sealed class SelectionSettings
@@ -7,4 +9,6 @@ internal sealed class SelectionSettings
     public List<SelectionHistoryEntry> History { get; set; } = new();
 
     public int RandomPosition { get; set; }
+
+    public Dictionary<uint, GameUserData> UserData { get; set; } = new();
 }